@@ -3,8 +3,8 @@ import CheckinsHistory from "./CheckinHistory";
 import { AppContextProvider } from "../../context/AppContext";
 import { Router } from "react-router-dom";
 import { createBrowserHistory } from "history";
-import { LocalizationProvider } from "@mui/lab";
-import AdapterDateFns from "@mui/lab/AdapterDateFns";
+import { LocalizationProvider } from "@mui/x-date-pickers";
+import { AdapterDateFns } from '@mui/x-date-pickers/AdapterDateFns';
 
 const mockMemberId = "bf9975f8-a5b2-4551-b729-afd56b49e2cc";
 const mockCheckinId = "3a1906df-d45c-4ff5-a6f8-7dacba97ff1a";
